Moved code from Json to DeriveSchema and removed jsonUtils
